refactor: change method name
