+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/progressmission/business/MissionBusiness.java
@@ -1,7 +1,7 @@
 package com.javabom.bomplatform.web.progressmission.business;
 
-import com.javabom.bomplatform.mission.model.Mission;
-import com.javabom.bomplatform.mission.repository.MissionRepository;
+import com.javabom.bomplatform.core.mission.model.Mission;
+import com.javabom.bomplatform.core.mission.repository.MissionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ public class MissionBusiness {
 
     private final MissionRepository missionRepository;
 
-    public Mission findById(Long missionId) {
+    public Mission showDetailById(Long missionId) {
         return missionRepository.findById(missionId)
                 .orElseThrow(() -> new NoSuchElementException(String.format("input id: %d, not found mission", missionId)));
     }
