+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/progressmission/business/ProgressMissionBusiness.java
@@ -1,11 +1,11 @@
 package com.javabom.bomplatform.web.progressmission.business;
 
-import com.javabom.bomplatform.mission.model.Mission;
-import com.javabom.bomplatform.progressmission.model.MissionReviewer;
-import com.javabom.bomplatform.progressmission.model.ProgressMission;
-import com.javabom.bomplatform.progressmission.model.ProgressMissionState;
-import com.javabom.bomplatform.progressmission.repository.ProgressMissionRepository;
-import com.javabom.bomplatform.user.model.User;
+import com.javabom.bomplatform.core.mission.model.Mission;
+import com.javabom.bomplatform.core.progressmission.model.MissionReviewer;
+import com.javabom.bomplatform.core.progressmission.model.ProgressMission;
+import com.javabom.bomplatform.core.progressmission.model.ProgressMissionState;
+import com.javabom.bomplatform.core.progressmission.repository.ProgressMissionRepository;
+import com.javabom.bomplatform.core.user.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,8 +19,6 @@ public class ProgressMissionBusiness {
 
     private final ProgressMissionRepository progressMissionRepository;
 
-    // naming (시작을 하는데 ProgressMission을 만든다고 예측하기 어렵다고 판단
-    // 후보: register
     public Long begin(Mission mission, MissionReviewer missionReviewer, User challenger) {
         final ProgressMission progressMission = ProgressMission.builder()
                 .mission(mission)
@@ -33,7 +31,7 @@ public class ProgressMissionBusiness {
         return progressMission.getId();
     }
 
-    public ProgressMission findById(Long progressMissionId) {
+    public ProgressMission showDetailById(Long progressMissionId) {
         return progressMissionRepository.findById(progressMissionId)
                 .orElseThrow(() -> new NoSuchElementException(String.format("input id: %d, not found progressMission", progressMissionId)));
     }
