+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/progressmission/business/UserBusiness.java
@@ -1,7 +1,7 @@
 package com.javabom.bomplatform.web.progressmission.business;
 
-import com.javabom.bomplatform.user.model.User;
-import com.javabom.bomplatform.user.repository.UserRepository;
+import com.javabom.bomplatform.core.user.model.User;
+import com.javabom.bomplatform.core.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,7 +13,7 @@ public class UserBusiness {
 
     private final UserRepository userRepository;
 
-    public User findChallenger(Long userId) {
+    public User findChallengerById(Long userId) {
         final User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NoSuchElementException(String.format("input id: %d, not found user", userId)));
