+++ bomplatform-web/src/main/java/com/javabom/bomplatform/web/review/service/ReviewService.java
@@ -1,12 +1,14 @@
 package com.javabom.bomplatform.web.review.service;
 
-import com.javabom.bomplatform.progressmission.model.ProgressMission;
-import com.javabom.bomplatform.web.github.service.GithubService;
+import com.javabom.bomplatform.core.progressmission.model.ProgressMission;
+import com.javabom.bomplatform.github.service.GithubService;
 import com.javabom.bomplatform.web.progressmission.business.ProgressMissionBusiness;
 import com.javabom.bomplatform.web.review.business.ReviewBusiness;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Arrays;
 
 @Service
 @RequiredArgsConstructor
@@ -19,9 +21,9 @@ public class ReviewService {
 
     @Transactional
     public Long request(Long progressMissionId, String reviewURL) {
-        final ProgressMission progressMission = progressMissionBusiness.findById(progressMissionId);
+        final ProgressMission progressMission = progressMissionBusiness.showDetailById(progressMissionId);
 
-        githubService.assignReviewer(reviewURL, progressMission.getMissionReviewers());
+        githubService.assignReviewer(reviewURL, Arrays.asList("missionReviewers"));
 
         return reviewBusiness.requestReview(progressMission, reviewURL);
     }
